Register mixin configs manually instead of using the interface
The only thing that really required using MixinBooter specifically was that interface. Since we have a coremod already anyway (and it probably will have to stay), we can just register the mixins in there instead of having MixinBooter do it. This lets users use whatever Mixin provider they want (as long as it meets the requirements).