Revert "Register mixin configs manually instead of using the interface"
This reverts commit 2b0e3f664059ef561fa6fcb490e50e58f546cd13.

Apparently "Late Mixins" are a hack on 1.12 and before due to when non-coremod mod classes are loaded. That means I either have to use a library, or maintain a bunch of hacks myself (the "best" of which seem to involve reflection against a moving target to reload mixin configs), so I'd rather not maintain that myself.