Fix research table shift click into table not working (#99)
For some reason, Mixin kept claiming it couldn't find an obfuscation mapping for isItemValidForSlot. It's definitely obfuscated, so I have no idea why it is confused and remap = false is not correct. Maybe it doesn't like that the method lives in a non-MC class, but overrides a method from an obfuscated interface?